Moved configuration update function to new file.
